External name check for wildcard patterns
- makeWildCardPattern moved to BasicUtils
- Checks for * used as a wildcard or as part of externalName
